Updated readme.md
From using the program and reviewing the code, it looks like Windows payloads must be "reverse_shell_tcp_inline" or other options, not "reverse_shell_tcp". I've made very small change to readme.md to fix this. Of course feel free to reject if this is too small or not accurate.